refactor(core): move signals package to pkg
